refactor: split and move tests
