+++ crates/iceberg/src/spec/transform.rs
@@ -666,11 +666,6 @@ enum AdjustedProjection {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use std::collections::HashSet;
-    use std::sync::Arc;
-
-    use crate::expr::{BoundPredicate, BoundReference, PredicateOperator};
     use crate::spec::datatypes::PrimitiveType::{
         Binary, Date, Decimal, Fixed, Int, Long, String as StringType, Time, Timestamp,
         Timestamptz, Uuid,
@@ -678,7 +673,6 @@ mod tests {
     use crate::spec::datatypes::Type::{Primitive, Struct};
     use crate::spec::datatypes::{NestedField, StructType, Type};
     use crate::spec::transform::Transform;
-    use crate::spec::{Datum, NestedFieldRef, PrimitiveType};
 
     struct TestParameter {
         display: String,
@@ -687,64 +681,6 @@ mod tests {
         preserves_order: bool,
         satisfies_order_of: Vec<(Transform, bool)>,
         trans_types: Vec<(Type, Option<Type>)>,
-    }
-
-    struct TestProjectionParameter {
-        transform: Transform,
-        name: String,
-        field: NestedFieldRef,
-    }
-
-    impl TestProjectionParameter {
-        fn new(transform: Transform, name: impl Into<String>, field: NestedField) -> Self {
-            TestProjectionParameter {
-                transform,
-                name: name.into(),
-                field: Arc::new(field),
-            }
-        }
-        fn name(&self) -> String {
-            self.name.clone()
-        }
-        fn field(&self) -> NestedFieldRef {
-            self.field.clone()
-        }
-        fn project(&self, predicate: &BoundPredicate) -> Result<Option<Predicate>> {
-            self.transform.project(self.name(), predicate)
-        }
-        fn _unary_predicate(&self, op: PredicateOperator) -> BoundPredicate {
-            BoundPredicate::Unary(UnaryExpression::new(
-                op,
-                BoundReference::new(self.name(), self.field()),
-            ))
-        }
-        fn binary_predicate(&self, op: PredicateOperator, literal: Datum) -> BoundPredicate {
-            BoundPredicate::Binary(BinaryExpression::new(
-                op,
-                BoundReference::new(self.name(), self.field()),
-                literal,
-            ))
-        }
-        fn set_predicate(&self, op: PredicateOperator, literals: Vec<Datum>) -> BoundPredicate {
-            BoundPredicate::Set(SetExpression::new(
-                op,
-                BoundReference::new(self.name(), self.field()),
-                HashSet::from_iter(literals),
-            ))
-        }
-    }
-
-    fn assert_projection(
-        predicate: &BoundPredicate,
-        fixture: &TestProjectionParameter,
-        expected: Option<&str>,
-    ) -> Result<()> {
-        let result = fixture.project(predicate)?;
-        match expected {
-            Some(exp) => assert_eq!(format!("{}", result.unwrap()), exp),
-            None => assert!(result.is_none()),
-        }
-        Ok(())
     }
 
     fn check_transform(trans: Transform, param: TestParameter) {
@@ -768,2937 +704,6 @@ mod tests {
         for (input_type, result_type) in param.trans_types {
             assert_eq!(result_type, trans.result_type(&input_type).ok());
         }
-    }
-
-    #[test]
-    fn test_projection_timestamp_hour_upper_bound() -> Result<()> {
-        // 420034
-        let value = "2017-12-01T10:59:59.999999";
-        // 412007
-        let another = "2016-12-31T23:59:59.999999";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Hour,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 420034"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 420034"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 420035"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 420034"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name = 420034"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (420034, 412007)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_hour_lower_bound() -> Result<()> {
-        // 420034
-        let value = "2017-12-01T10:00:00.000000";
-        // 411288
-        let another = "2016-12-02T00:00:00.000000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Hour,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 420033"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 420034"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 420034"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 420034"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name = 420034"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (420034, 411288)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_year_upper_bound() -> Result<()> {
-        let value = "2017-12-31T23:59:59.999999";
-        let another = "2016-12-31T23:59:59.999999";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Year,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 48"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name = 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (47, 46)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_year_lower_bound() -> Result<()> {
-        let value = "2017-01-01T00:00:00.000000";
-        let another = "2016-12-02T00:00:00.000000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Year,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 46"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name = 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (47, 46)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_month_negative_upper_bound() -> Result<()> {
-        let value = "1969-12-31T23:59:59.999999";
-        let another = "1970-01-01T00:00:00.000000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Month,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= -1"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name IN (-1, 0)"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (0, -1)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_month_upper_bound() -> Result<()> {
-        let value = "2017-12-01T23:59:59.999999";
-        let another = "2017-11-02T00:00:00.000000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Month,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name = 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (575, 574)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_month_negative_lower_bound() -> Result<()> {
-        let value = "1969-01-01T00:00:00.000000";
-        let another = "1969-03-01T00:00:00.000000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Month,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= -12"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= -11"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= -12"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= -12"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name IN (-12, -11)"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (-10, -9, -12, -11)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_month_lower_bound() -> Result<()> {
-        let value = "2017-12-01T00:00:00.000000";
-        let another = "2017-12-02T00:00:00.000000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Month,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 574"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name = 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (575)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_day_negative_upper_bound() -> Result<()> {
-        // -1
-        let value = "1969-12-31T23:59:59.999999";
-        // 0
-        let another = "1970-01-01T00:00:00.000000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Day,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= -1"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name IN (-1, 0)"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (0, -1)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_day_upper_bound() -> Result<()> {
-        // 17501
-        let value = "2017-12-01T23:59:59.999999";
-        // 17502
-        let another = "2017-12-02T00:00:00.000000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Day,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 17501"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 17501"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 17502"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 17501"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name = 17501"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (17501, 17502)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_day_negative_lower_bound() -> Result<()> {
-        // -365
-        let value = "1969-01-01T00:00:00.000000";
-        // -364
-        let another = "1969-01-02T00:00:00.000000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Day,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= -365"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= -364"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= -365"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= -365"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name IN (-364, -365)"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (-363, -364, -365)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_day_lower_bound() -> Result<()> {
-        // 17501
-        let value = "2017-12-01T00:00:00.000000";
-        // 17502
-        let another = "2017-12-02T00:00:00.000000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Day,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 17500"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 17501"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 17501"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 17501"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name = 17501"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (17501, 17502)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_timestamp_day_epoch() -> Result<()> {
-        // 0
-        let value = "1970-01-01T00:00:00.00000";
-        // 1
-        let another = "1970-01-02T00:00:00.00000";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Day,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Timestamp)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::timestamp_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            Some("name = 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::timestamp_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (1, 0)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::timestamp_from_str(value)?,
-                    Datum::timestamp_from_str(another)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_day_negative() -> Result<()> {
-        // -2
-        let value = "1969-12-30";
-        // -4
-        let another = "1969-12-28";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Day,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= -3"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= -2"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= -1"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= -2"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name = -2"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (-2, -4)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_day() -> Result<()> {
-        // 17167
-        let value = "2017-01-01";
-        // 17531
-        let another = "2017-12-31";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Day,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= 17166"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 17167"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= 17168"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 17167"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name = 17167"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (17531, 17167)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_month_negative_upper_bound() -> Result<()> {
-        // -1 => 1969-12
-        let value = "1969-12-31";
-        // -12 => 1969-01
-        let another = "1969-01-01";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Month,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= -1"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name IN (-1, 0)"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (-1, -12, -11, 0)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_month_upper_bound() -> Result<()> {
-        // 575 => 2017-12
-        let value = "2017-12-31";
-        // 564 => 2017-01
-        let another = "2017-01-01";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Month,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= 576"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name = 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (575, 564)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_month_negative_lower_bound() -> Result<()> {
-        // -12 => 1969-01
-        let value = "1969-01-01";
-        // -1 => 1969-12
-        let another = "1969-12-31";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Month,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= -12"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= -11"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= -12"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= -12"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name IN (-12, -11)"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (-1, -12, -11, 0)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_month_lower_bound() -> Result<()> {
-        // 575 => 2017-12
-        let value = "2017-12-01";
-        // 564 => 2017-01
-        let another = "2017-01-01";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Month,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= 574"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name = 575"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (575, 564)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_month_epoch() -> Result<()> {
-        // 0 => 1970-01
-        let value = "1970-01-01";
-        // -1 => 1969-12
-        let another = "1969-12-31";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Month,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name = 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (0, -1)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_year_negative_upper_bound() -> Result<()> {
-        // -1 => 1969
-        let value = "1969-12-31";
-        let another = "1969-01-01";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Year,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= -1"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name IN (-1, 0)"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (0, -1)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_year_upper_bound() -> Result<()> {
-        // 47 => 2017
-        let value = "2017-12-31";
-        // 46 => 2016
-        let another = "2016-01-01";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Year,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= 48"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name = 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (47, 46)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_year_negative_lower_bound() -> Result<()> {
-        // 0 => 1970
-        let value = "1970-01-01";
-        // -1 => 1969
-        let another = "1969-12-31";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Year,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name = 0"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (0, -1)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_date_year_lower_bound() -> Result<()> {
-        // 47 => 2017
-        let value = "2017-01-01";
-        // 46 => 2016
-        let another = "2016-12-31";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Year,
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Date)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name <= 46"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name <= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name >= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::date_from_str(value)?,
-            ),
-            &fixture,
-            Some("name >= 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::date_from_str(value)?),
-            &fixture,
-            Some("name = 47"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::date_from_str(value)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            Some("name IN (47, 46)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::date_from_str(value)?, Datum::date_from_str(another)?],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_truncate_string_rewrite_op() -> Result<()> {
-        let fixture = TestProjectionParameter::new(
-            Transform::Truncate(5),
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::String)),
-        );
-
-        let value = "abcde";
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::StartsWith, Datum::string(value)),
-            &fixture,
-            Some(r#"name = "abcde""#),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotStartsWith, Datum::string(value)),
-            &fixture,
-            Some(r#"name != "abcde""#),
-        )?;
-
-        let value = "abcdefg";
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::StartsWith, Datum::string(value)),
-            &fixture,
-            Some(r#"name STARTS WITH "abcde""#),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotStartsWith, Datum::string(value)),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_truncate_string() -> Result<()> {
-        let value = "abcdefg";
-        let fixture = TestProjectionParameter::new(
-            Transform::Truncate(5),
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::String)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::string(value)),
-            &fixture,
-            Some(r#"name <= "abcde""#),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThanOrEq, Datum::string(value)),
-            &fixture,
-            Some(r#"name <= "abcde""#),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::string(value)),
-            &fixture,
-            Some(r#"name >= "abcde""#),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThanOrEq, Datum::string(value)),
-            &fixture,
-            Some(r#"name >= "abcde""#),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::string(value)),
-            &fixture,
-            Some(r#"name = "abcde""#),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::string(value), Datum::string(format!("{}abc", value))],
-            ),
-            &fixture,
-            Some(r#"name IN ("abcde")"#),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::string(value), Datum::string(format!("{}abc", value))],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_truncate_upper_bound_decimal() -> Result<()> {
-        let prev = "98.99";
-        let curr = "99.99";
-        let next = "100.99";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Truncate(10),
-            "name",
-            NestedField::required(
-                1,
-                "value",
-                Type::Primitive(PrimitiveType::Decimal {
-                    precision: 9,
-                    scale: 2,
-                }),
-            ),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::decimal_from_str(curr)?),
-            &fixture,
-            Some("name <= 9990"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::decimal_from_str(curr)?,
-            ),
-            &fixture,
-            Some("name <= 9990"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::decimal_from_str(curr)?,
-            ),
-            &fixture,
-            Some("name >= 9990"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::decimal_from_str(curr)?),
-            &fixture,
-            Some("name = 9990"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::decimal_from_str(curr)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::decimal_from_str(prev)?,
-                    Datum::decimal_from_str(curr)?,
-                    Datum::decimal_from_str(next)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (10090, 9990, 9890)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::decimal_from_str(curr)?,
-                    Datum::decimal_from_str(next)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_truncate_lower_bound_decimal() -> Result<()> {
-        let prev = "99.00";
-        let curr = "100.00";
-        let next = "101.00";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Truncate(10),
-            "name",
-            NestedField::required(
-                1,
-                "value",
-                Type::Primitive(PrimitiveType::Decimal {
-                    precision: 9,
-                    scale: 2,
-                }),
-            ),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::decimal_from_str(curr)?),
-            &fixture,
-            Some("name <= 9990"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::decimal_from_str(curr)?,
-            ),
-            &fixture,
-            Some("name <= 10000"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::decimal_from_str(curr)?,
-            ),
-            &fixture,
-            Some("name >= 10000"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::decimal_from_str(curr)?),
-            &fixture,
-            Some("name = 10000"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::decimal_from_str(curr)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::decimal_from_str(prev)?,
-                    Datum::decimal_from_str(curr)?,
-                    Datum::decimal_from_str(next)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (9900, 10000, 10100)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::decimal_from_str(curr)?,
-                    Datum::decimal_from_str(next)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_truncate_upper_bound_long() -> Result<()> {
-        let value = 99i64;
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Truncate(10),
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Long)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::long(value)),
-            &fixture,
-            Some("name <= 90"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThanOrEq, Datum::long(value)),
-            &fixture,
-            Some("name <= 90"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThanOrEq, Datum::long(value)),
-            &fixture,
-            Some("name >= 90"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::long(value)),
-            &fixture,
-            Some("name = 90"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::long(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::long(value - 1),
-                    Datum::long(value),
-                    Datum::long(value + 1),
-                ],
-            ),
-            &fixture,
-            Some("name IN (100, 90)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::long(value), Datum::long(value + 1)],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_truncate_lower_bound_long() -> Result<()> {
-        let value = 100i64;
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Truncate(10),
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Long)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::long(value)),
-            &fixture,
-            Some("name <= 90"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThanOrEq, Datum::long(value)),
-            &fixture,
-            Some("name <= 100"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThanOrEq, Datum::long(value)),
-            &fixture,
-            Some("name >= 100"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::long(value)),
-            &fixture,
-            Some("name = 100"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::long(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::long(value - 1),
-                    Datum::long(value),
-                    Datum::long(value + 1),
-                ],
-            ),
-            &fixture,
-            Some("name IN (100, 90)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::long(value), Datum::long(value + 1)],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_truncate_upper_bound_integer() -> Result<()> {
-        let value = 99;
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Truncate(10),
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Int)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::int(value)),
-            &fixture,
-            Some("name <= 90"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThanOrEq, Datum::int(value)),
-            &fixture,
-            Some("name <= 90"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThanOrEq, Datum::int(value)),
-            &fixture,
-            Some("name >= 90"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::int(value)),
-            &fixture,
-            Some("name = 90"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::int(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::int(value - 1),
-                    Datum::int(value),
-                    Datum::int(value + 1),
-                ],
-            ),
-            &fixture,
-            Some("name IN (100, 90)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::int(value), Datum::int(value + 1)],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_truncate_lower_bound_integer() -> Result<()> {
-        let value = 100;
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Truncate(10),
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Int)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::int(value)),
-            &fixture,
-            Some("name <= 90"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThanOrEq, Datum::int(value)),
-            &fixture,
-            Some("name <= 100"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThanOrEq, Datum::int(value)),
-            &fixture,
-            Some("name >= 100"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::int(value)),
-            &fixture,
-            Some("name = 100"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::int(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::int(value - 1),
-                    Datum::int(value),
-                    Datum::int(value + 1),
-                ],
-            ),
-            &fixture,
-            Some("name IN (100, 90)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::int(value), Datum::int(value + 1)],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_bucket_uuid() -> Result<()> {
-        let value = uuid::Uuid::from_u64_pair(123, 456);
-        let another = uuid::Uuid::from_u64_pair(456, 123);
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Bucket(10),
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Uuid)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::uuid(value)),
-            &fixture,
-            Some("name = 4"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::uuid(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::uuid(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThanOrEq, Datum::uuid(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::uuid(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThanOrEq, Datum::uuid(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::uuid(value), Datum::uuid(another)],
-            ),
-            &fixture,
-            Some("name IN (4, 6)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::uuid(value), Datum::uuid(another)],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_bucket_fixed() -> Result<()> {
-        let value = "abcdefg".as_bytes().to_vec();
-        let another = "abcdehij".as_bytes().to_vec();
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Bucket(10),
-            "name",
-            NestedField::required(
-                1,
-                "value",
-                Type::Primitive(PrimitiveType::Fixed(value.len() as u64)),
-            ),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::fixed(value.clone())),
-            &fixture,
-            Some("name = 4"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::fixed(value.clone())),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::fixed(value.clone())),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThanOrEq, Datum::fixed(value.clone())),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::fixed(value.clone())),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::fixed(value.clone()),
-            ),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::fixed(value.clone()), Datum::fixed(another.clone())],
-            ),
-            &fixture,
-            Some("name IN (4, 6)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::fixed(value.clone()), Datum::fixed(another.clone())],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_bucket_string() -> Result<()> {
-        let value = "abcdefg";
-        let another = "abcdefgabc";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Bucket(10),
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::String)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::string(value)),
-            &fixture,
-            Some("name = 4"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::string(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::string(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThanOrEq, Datum::string(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::string(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThanOrEq, Datum::string(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![Datum::string(value), Datum::string(another)],
-            ),
-            &fixture,
-            Some("name IN (9, 4)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::string(value), Datum::string(another)],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_bucket_decimal() -> Result<()> {
-        let prev = "99.00";
-        let curr = "100.00";
-        let next = "101.00";
-
-        let fixture = TestProjectionParameter::new(
-            Transform::Bucket(10),
-            "name",
-            NestedField::required(
-                1,
-                "value",
-                Type::Primitive(PrimitiveType::Decimal {
-                    precision: 9,
-                    scale: 2,
-                }),
-            ),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::decimal_from_str(curr)?),
-            &fixture,
-            Some("name = 2"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::decimal_from_str(curr)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::decimal_from_str(curr)?),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::LessThanOrEq,
-                Datum::decimal_from_str(curr)?,
-            ),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThan,
-                Datum::decimal_from_str(curr)?,
-            ),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(
-                PredicateOperator::GreaterThanOrEq,
-                Datum::decimal_from_str(curr)?,
-            ),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::decimal_from_str(next)?,
-                    Datum::decimal_from_str(curr)?,
-                    Datum::decimal_from_str(prev)?,
-                ],
-            ),
-            &fixture,
-            Some("name IN (6, 2)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![
-                    Datum::decimal_from_str(curr)?,
-                    Datum::decimal_from_str(next)?,
-                ],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_bucket_long() -> Result<()> {
-        let value = 100;
-        let fixture = TestProjectionParameter::new(
-            Transform::Bucket(10),
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Long)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::long(value)),
-            &fixture,
-            Some("name = 6"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::long(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::long(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThanOrEq, Datum::long(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::long(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThanOrEq, Datum::long(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::long(value - 1),
-                    Datum::long(value),
-                    Datum::long(value + 1),
-                ],
-            ),
-            &fixture,
-            Some("name IN (8, 7, 6)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::long(value), Datum::long(value + 1)],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_projection_bucket_integer() -> Result<()> {
-        let value = 100;
-        let fixture = TestProjectionParameter::new(
-            Transform::Bucket(10),
-            "name",
-            NestedField::required(1, "value", Type::Primitive(PrimitiveType::Int)),
-        );
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::Eq, Datum::int(value)),
-            &fixture,
-            Some("name = 6"),
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::NotEq, Datum::int(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThan, Datum::int(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::LessThanOrEq, Datum::int(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThan, Datum::int(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.binary_predicate(PredicateOperator::GreaterThanOrEq, Datum::int(value)),
-            &fixture,
-            None,
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::In,
-                vec![
-                    Datum::int(value - 1),
-                    Datum::int(value),
-                    Datum::int(value + 1),
-                ],
-            ),
-            &fixture,
-            Some("name IN (8, 7, 6)"),
-        )?;
-
-        assert_projection(
-            &fixture.set_predicate(
-                PredicateOperator::NotIn,
-                vec![Datum::int(value), Datum::int(value + 1)],
-            ),
-            &fixture,
-            None,
-        )?;
-
-        Ok(())
     }
 
     #[test]
